requirements.txt
# Core data science libraries
pandas>=2.0.0
numpy>=1.24.0

# Web framework
streamlit>=1.28.0

# Machine learning
scikit-learn>=1.3.0
scipy>=1.11.0

# Data visualization
matplotlib>=3.7.0
plotly>=5.15.0
seaborn>=0.12.0
plotly>=5.15.0

# Jupyter and notebook environment
jupyter>=1.0.0
notebook>=6.5.0
ipykernel>=6.24.0
ipywidgets>=8.0.0

# GUI libraries
pillow>=10.0.0

# Web development (optional)
flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0

# Additional utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
joblib>=1.3.0
python-dotenv>=1.0.0

# Model persistence
joblib>=1.3.0

# Data processing
regex>=2023.6.3
